requirements.txt
# #Django==4.0.3
# Pillow
# django-qr-code

Django==3.2.22
Pillow==11.3.0
django-qr-code==3.1.1
django-cryptography==1.1
pydantic==2.12.3
pydantic-core==2.41.4
